move shape padding back to partial derivatives
